feat(gost): Support Debian Security Tracker
